Extract MetricStore into its own module
